% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{cordes_npm_install}
\alias{cordes_npm_install}
\title{Install Npm infra inside Cordes Package}
\usage{
cordes_npm_install(
  package,
  force = FALSE,
 
    yes_no_message = "This will install plop npm dependencies on your local library. \\nAre you ok with that?",
  folder = "cordes"
)
}
\arguments{
\item{package}{package name}

\item{force}{Prompt the user to accept install?}

\item{yes_no_message}{What to display as a prompt message}

\item{folder}{Where to got to run \verb{npm install}}
}
\description{
Install Npm infra inside Cordes Package
}
