% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{cordes_run}
\alias{cordes_run}
\title{Run Node Script}
\usage{
cordes_run(package, extra_args = c(), file = "app.js", folder = "cordes")
}
\arguments{
\item{package}{package name}

\item{folder}{Where to got to run \verb{npm install}}
}
\description{
Run Node Script
}
