% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm_add_module.R
\name{install_module}
\alias{install_module}
\title{Add a module to the Corde Package}
\usage{
install_module(module, npm_opt = c())
}
\arguments{
\item{module}{module name}

\item{npm_opt}{Vector of options passed to \verb{npm install}}
}
\description{
Add a module to the Corde Package
}
