% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{cordes_npm_install}
\alias{cordes_npm_install}
\title{Title}
\usage{
cordes_npm_install(
  package,
  force = FALSE,
 
    yes_no_message = "This will install plop npm dependencies on your local library. \\nAre you ok with that?",
  folder = "cordes"
)
}
\arguments{
\item{yes_no_message}{}
}
\value{

}
\description{
Title
}
